Enum query_projector_traits::aggregates::SimpleAggregationOp

pub enum SimpleAggregationOp {
    Avg,
    Count,
    Max,
    Min,
    Sum,
}

Variants
AvgCountMaxMinSum
Methods

impl SimpleAggregationOp

pub fn to_sql(&self) -> &'static str

pub fn is_applicable_to_types(
&self,
possibilities: ValueTypeSet
) -> Result<ValueType>

With knowledge of the types to which a variable might be bound, return a Result to determine whether this aggregation is suitable. For example, it's valid to take the Avg of {Double, Long}, invalid to take Sum of {Instant}, valid to take (lexicographic) Max of {String}, but invalid to take Max of {Uuid, String}.

The returned type is the type of the result of the aggregation.